Cleanup ConfigureView
